// apps/api/src/domains/review-approval/domain/workflow-status.ts
export enum ProjectWorkflowStatus {
  DRAFT = 'DRAFT',
  READY_FOR_REVIEW = 'READY_FOR_REVIEW',
  IN_REVIEW = 'IN_REVIEW',
  CHANGES_REQUESTED = 'CHANGES_REQUESTED',
  RESUBMITTED = 'RESUBMITTED',
  APPROVED = 'APPROVED',
  ARCHIVED = 'ARCHIVED',
  REJECTED = 'REJECTED',
  CANCELLED = 'CANCELLED',
}

export enum SectionWorkflowStatus {
  DRAFT = 'DRAFT',
  COMPLETE = 'COMPLETE',
  IN_REVIEW = 'IN_REVIEW',
  CHANGES_REQUESTED = 'CHANGES_REQUESTED',
  APPROVED = 'APPROVED',
}